added Landables Value option MaxOkaySink
This adds the subject new option on config menu 11 (Appearance).
The online help is modified accordingly, and "Lift" on config
menu 9 (Units) is changed to "Lift/Sink".  If this option is
selected, instead of arrival altitude or glide ratio to target,
the max avg amount of sink that could be found on glide to the
target and arrive there at exactly the safety altitude is shown.