Optimized routes for paragliders. Draft, but fully working. Leaning on AAT locked target and virtual waypoints, a pretty effective result. THANKS to Donato, DigiFly Europe, for some important parts of the code, help, and support!
Optimization can be enabled in Paragliders config menu 23, and as a customkey.
When enabled, the T> overlay, the Next waypoint distance, altitude diff, bearing etc.
are all related to the optimized turn point on the cylinder.
The correct optimized point of turn is visualized also.
The optimized turnpoint is named with a "!" leading. Example: turnpoint name is "JACKHILL"
as optimized point it will appear as T>!JACKHILL   and also in turnpoint list and nearest pages.
This is a virtual waypoint, like thermal, takeoff etc.

Since Optimized Route (O.R.) rely on the internal AAT engine for gliders, it is better that
paragliders check the TARGET  cylinder radius to be at -100% , in order to let the calculator
consider the max radius as a turnpoint. By default, but only when you create a new task as
paraglider with this version, all radius are set to -100%. To check, menu NAV 2/3 and then
"Target" , then select turnpoints exluding first and last (start and finish).
Check that Range is -100%, and Radial is 0.