New infoboxes for Max Altitude reached, Max Height Gained, Head Wind Speed.
[Max Altitude Reached]
Max altitude reached during the current flight.
It is reset at takeoff and at task restart.

[Max Height Gained]
This height is calculated using the min and max heights detected AFTER the start of free flight.
This value is normally used for badges.

[Head Wind Speed]
This is the wind component relative to your current track.
It is the difference between the TAS true airspeed and GS ground speed.
A positive value indicates a HEAD WIND, and a negative value indicates a TAIL WIND.
If true airspeed is not available, estimated TAS is used.
In this case the value is only guessed against the current wind set.