Decrease airspace warnlevel to yellow, when leaving a nonfly zone
This was missing, because the original approach was to not give repeated
red warnings when circling on an asp border. However the right behavior
should be to give warnings, and the user has to ack it if he doesn't want
to see that warning. It is also important if flying near an asp border in
parallel with it.