. Remove ProjectedTrack dash line for AAT tasks. + . Add "Position" button in SIM mode PAN mode.
To reposition the aircraft in sim pan mode, simply change position by dragging the moving map, and when the center cross is
where you want to position the aircraft, click on the "Position" new button on left top screen.
The old behaviour was to long click on the moving map, but it was not working fine on PNAs, and it was inaccurate.
Now that we have a smooth scrolling, it is better to use a button.
The old function for repositioning is no more available.
XCI has changed, update DEFAULT_MENU