NEWMULTIMAPS , in options.h reversible for a while.. Inside Multimap.cpp we decide which multimap page is "custom" or "shared". If it is custom, it is passed EVENTS_TOPLEFT TOPRIGHT SHORTCLICK, and the function should manage manually each keypress (UP, DOWN.. included). Like for current old multimaps, for RADAR, for TEST etc.
If it is a shared multimap, LK will use the same events and customkeys as for the main map:
customkeys are in use, overlays are printed (if activated) etc.
There are no SHORTCLICK, TOPLEFT and TOPRIGHT, because there are no such events in main map.
TOPLEFT and TOPRIGHT in fact are used by customkeys.

Bottombar is standalone, in any case.
Extensive tests needed for paragliding mode.
This is a preliminary, but hopefully usable, version.