Fix Zoom speed, zoom buttons, and topology refresh. ForceRenderMap was badly used in Thread_Draw to force topology scan (through UpdateCache forced) everytime it was set in zoom. This was greatly slowing down performances. We put back the first_run flag for the Thread_Draw, and keep at the same time the ForceRenderMap flag for zooms.
Zoom buttons (not zoom customkeys) were not triggering BigZoom and thus they were delayed.
Still we dont lower debounce time for them, like we do for map clicks. But it is a really minor issue.

VisibilityScan was not set after panning mode off. This may not be a good solution for all the situations,
and we might use it only for SIM mode and not for flights, so further checks are required.

Took several hours to check these things, because all checks must be made on real PNAs, since the PC is too fast and we probably gain a loop run (or more) in thread_draw , catching the next click withouth having latencies like on slower devices.