Target dialog for AAT sector fixed
Before the change a Target dialog for narrow AAT sector (like 90o - 180o) was
centered at the center of the circle used to draw the sector and zoomed
to view 2x its radius. Because of that sector was really hard to manage in
AAT tasks. Now the same sector is centered on the screen and stretched on full
screen.