[XCS patch] Android/TopWindow: fix deadlock while stopping XCSoar
Quite a few ANRs have been reported over the years with XCSoar hanging
in TopWindow::Pause().  In these ANR reports, there was no native
thread, which suggests that XCSoar has been quitting, and
concurrently, onPause() has been called.  So TopWindow::Pause()
generated a `PAUSE` event and waited for it to be handled; but there
was nobody anymore the the other end of the EventQueue.