Show black circle for static Flarm-Objects
Changed icons for static Flarm-Object (on ground-pilots) to black circle.
A static object (ground-object) is determined, when speed is 0
according to Flarm data port specification

<GroundSpeed> : Decimal integer value. Range: from 0 to 32767.
When the aircraft is considered moving, the target’s ground
speed in m/s, forced to > 0.
When the aircraft is considered on the ground, the field is forced to 0.
This field is empty if stealth mode is activated either on the target
or own aircraft and for non-directional target

Manually merge Pull-Request #1548